Load the shiny app

This function is used to launch the COVID-19 Global Analysis Shiny Application from the COVID19bb package.

launch_app()

Details

This app is intended to be used as an intuitive and easy-to-understand visualization tool for COVID-19 research, users can observe the total distribution of confirmed cases, death cases and recovered cases around the world by selecting the type of the distribution they want to view, as well as exploring the trends of mortality, diagnosis recovered counts among provinces or states of different countries all over the world.

Examples

if (FALSE) {
launch_app()
}